# swagger.yaml
openapi: 3.0.3
info:
  title: AXZY PARKING API
  description: API para control de estacionamiento, entradas, salidas, movimientos y evidencia.
  version: 1.0.0
servers:
  - url: http://localhost:4444/api/v1
    description: Local server
...
paths:
...
  /kardex:
    post:
      tags: [kardex]
      summary: Crear registro de kardex
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KardexCreate'
      responses:
        "201":
           description: Le Created
           content:
             application/json:
               schema:
                 $ref: '#/components/schemas/TResultKardex'
    get:
      tags: [kardex]
      summary: Listar kardex (filtros opcionales)
      parameters:
        - in: query
          name: userId
          schema:
            type: integer
        - in: query
          name: locationId
          schema:
            type: integer
        - in: query
          name: startDate
          schema:
            type: string
            format: date-time
        - in: query
          name: endDate
          schema:
            type: string
            format: date-time
      responses:
         "200":
           description: OK
           content:
             application/json:
               schema:
                 $ref: '#/components/schemas/TResultKardexArray'
  /kardex/{id}:
    get:
      tags: [kardex]
      summary: Detalle de kardex
      parameters:
        - in: path
          name: id
          required: true
          schema: 
            type: integer
      responses:
         "200":
           description: OK
           content:
             application/json:
               schema:
                 $ref: '#/components/schemas/TResultKardex'




  ###########################################################################
  # SCHEDULES
  ###########################################################################
...
components:
  schemas:
...
    TResultKardex:
       type: object
       properties:
         data: { $ref: '#/components/schemas/Kardex' }
         error: { type: string, nullable: true }
    TResultKardexArray:
       type: object
       properties:
         data:
           type: array
           items: { $ref: '#/components/schemas/Kardex' }
         error: { type: string, nullable: true }



    ################################################################
    # ENTITIES
    ################################################################
...
    Kardex:
       type: object
       properties:
         id: { type: integer }
         userId: { type: integer }
         locationId: { type: integer }
         timestamp: { type: string, format: date-time }
         notes: { type: string }
         media: 
           type: array
           items:
             type: object
             properties:
                type: { type: string, enum: [IMAGE, VIDEO] }
                url: { type: string }

    ################################################################
    # PROPERTY SCHEMAS
    ################################################################
...
    KardexCreate:
       type: object
       required: [userId, locationId]
       properties:
         userId: { type: integer }
         locationId: { type: integer }
         notes: { type: string }
         media:
           type: array
           items:
             type: object
             properties:
               type: { type: string, enum: [IMAGE, VIDEO] }
               url: { type: string }